Revert "Backend & connection added"
