+++ src/pages/studentDashboard/StudentDashboard.jsx
@@ -2,6 +2,7 @@ import {
   Button,
   Card,
   CardActionArea,
+  CardActions,
   CardContent,
   CardMedia,
   Divider,
